Modularized domainDetails, so now it displays separated modules for different statuses
